[University] Extract tabs and improve query param handling
